requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
Django==4.0.5
django-bootstrap-datepicker-plus==4.0.0
django-bootstrap4==22.1
django-crispy-forms==1.14.0
install==1.3.5
soupsieve==2.3.2.post1
sqlparse==0.4.2


